Requests (requests v0.1.0-dev)

API

get(url, opts \\ [])

Makes a GET request.

See request/4 for possible options.

get!(url, opts \\ [])

See get/2.

request(method, url, body, opts \\ [])

Makes a HTTP request.

Options:

:headers - list of request headers, defaults to [].

:finch - name of the Finch pool to use, defaults to Requests.Finch that is started with the default options.

:request_middleware - list of middleware to run the request through, defaults to [].

:default_request_middleware - if true (default), prepends the following functions to the request middleware list:

normalize_request_headers/2

default_headers/2

:response_middleware - list of middleware to run the response through, defaults to [].

:default_response_middleware - if true (default), prepends the following functions to the response middleware list:

decompress/2

response_content_type/2

The opts keywords list is passed to each middleware.

Request middleware

A request middleware is a function that two arguments:

request struct

opts

An example is normalize_request_headers/2.

Response middleware

A response middleware is a function that two arguments:

response struct

opts

An example is decompress/2.

Request middleware

default_headers(request, opts)

Adds default request headers.

Currently these headers are added:

"user-agent" - "requests/0.1.0-dev"

"accept-encoding" - "gzip"

normalize_request_headers(request, opts)

Normalizes request headers.

Turns atom header names into strings, e.g.: :user_agent becomes "user-agent". Non-atom names are returned as is.

Response middleware

decompress(response, opts)

Decompresses the response body.

response_content_type(response, opts)

Decodes the response body based on the content-type header.

Options

:json_decoder - if set, used on the "application/json*" content type. Defaults to &Jason.decode/1 if jason dependency is installed.

:csv_decoder - if set, used on the "text/csv*" content type. Defaults to &NimbleCSV.RFC4180.parse_string(&1, skip_headers: false) if nimble_csv dependency is installed.