Requests (requests v0.1.0-dev)

Yet another HTTP client.

Features

Extensible via request and response middlewares

Automatic body encoding/decoding (via encode_request_body/2 and decode_response_body/2 middleware)

Automatic compression/decompression (via the compress/2 and decompress/2 middleware)

Examples

iex> Requests.get!("https://api.github.com/repos/elixir-lang/elixir").body["description"]
"Elixir is a dynamic, functional language designed for building scalable and maintainable applications"

iex> data = %{files: %{"hello.txt" => %{"content" => "world"}}}
iex> headers = [authorization: "Bearer " <> System.fetch_env!("GITHUB_TOKEN")]
iex> Requests.post!("https://api.github.com/gists", {:json, data}, headers: headers).status
201

Credits

Requests

Tesla

Finch (and Mint & NimblePool!)

API

get(url, opts \\ [])

Makes a GET request.

See request/4 for possible options.

get!(url, opts \\ [])

See get/2.

post(url, body, opts \\ [])

Makes a POST request.

See request/4 for possible options.

post!(url, body, opts \\ [])

See post/3.

request(method, url, body, opts \\ [])

Makes a HTTP request.

Options:

:headers - list of request headers, defaults to [].

:finch - name of the Finch pool to use, defaults to Requests.Finch that is started with the default options.

:request_middleware - list of middleware to run the request through, defaults to [].

:default_request_middleware - if true (default), prepends the following functions to the request middleware list:

normalize_request_headers/2

default_headers/2

encode_request_body/2

compress/2

:response_middleware - list of middleware to run the response through, defaults to [].

:default_response_middleware - if true (default), prepends the following functions to the response middleware list:

decompress/2

decode_response_body/2

The opts keywords list is passed to each middleware.

Request middleware

A request middleware is a function that takes two arguments: request:

a Finch.Request struct

an opts keywords list

and returns a possibly updated.

An example is compress/2.

Response middleware

A response middleware is a function that takes two arguments:

a Finch.Response struct

an opts keywords list

and returns a possibly updated response.

An example is decompress/2.

Request middleware

compress(request, opts)

Compresses the request body based on the content-encoding header.

Supported values: "gzip", "x-gzip", "deflate", and "identity".

default_headers(request, opts)

Adds default request headers.

Currently these headers are added:

"user-agent" - "requests/0.1.0-dev"

"accept-encoding" - "gzip"

encode_request_body(request, opts)

Encodes the request body based on its shape.

If body is of the following shape, it's encoded and its content-type set accordingly. Otherwise it's unchanged.

Shape	Encoder	Content-Type

{:form, data}	:form_encoder	"application/x-www-form-urlencoded"

{:json, data}	:json_encoder	"application/json"

{:csv, rows}	:csv_encoder	"text/csv"

{:csv, {:stream, rows}}	:csv_stream_encoder	"text/csv"

Options

:form_encoder - defaults to &URI.encode_query/1.

:json_encoder - defaults to &Jason.encode_to_iodata!/1.

:csv_encoder - defaults to &NimbleCSV.RFC4180.dump_to_iodata/1.

:csv_stream_encoder - defaults to &NimbleCSV.RFC4180.dump_to_stream/1.

Examples

iex> Requests.post!("https://httpbin.org/post", {:form, custname: "Alice"})
%Finch.Response{
  status: 200,
  headers: [{"content-type", "application/json"}, ...],
  body: %{
    "form" => %{"custname" => "Alice"},
    ...
  }
}

normalize_request_headers(request, opts)

Normalizes request headers.

Turns atom header names into strings, e.g.: :user_agent becomes "user-agent". Non-atom names are returned as is.

Response middleware

decode_response_body(response, opts)

Decodes the response body based on the content-type header.

Options

:json_decoder - if set, used on the "application/json*" content type. Defaults to &Jason.decode/1 if jason dependency is installed.

:csv_decoder - if set, used on the "text/csv*" content type. Defaults to &NimbleCSV.RFC4180.parse_string(&1, skip_headers: false) if nimble_csv dependency is installed.

decompress(response, opts)

Decompresses the response body based on the content-encoding header.

Supported values: "gzip", "x-gzip", "deflate", and "identity".