Requests (requests v0.1.0-dev)

Yet another HTTP client.

Features

Extensible via request, response, and error middleware

Automatic body encoding/decoding (via the encode/2 and decode/2 middleware)

Automatic compression/decompression (via the compress/2 and decompress/1 middleware)

Basic authentication (via the auth/2 middleware)

Retries on errors (via the retry/2 middleware)

Request streaming (by setting body as {:stream, enumerable})

Examples

iex> Requests.get!("https://api.github.com/repos/elixir-lang/elixir").body["description"]
"Elixir is a dynamic, functional language designed for building scalable and maintainable applications"

iex> Requests.post!("https://httpbin.org/post", {:form, comments: "hello!"}).body["form"]
%{"comments" => "hello!"}

Credits

Requests

Tesla

Finch (and Mint & NimblePool!)

API

auth(request, auth)

Sets request authentication.

auth can be one of:

{username, password} - same as {:basic, username, password}

{:basic, username, password} - uses Basic HTTP authentication

{:bearer, token} - sets OAuth 2.0 bearer token

Examples

iex> Requests.get!("https://httpbin.org/basic-auth/foo/bar", auth: {"bad", "bad"}).status
401
iex> Requests.get!("https://httpbin.org/basic-auth/foo/bar", auth: {"foo", "bar"}).status
200

get(url, opts \\ [])

Makes a GET request.

See request/4 for possible options.

get!(url, opts \\ [])

See get/2.

post(url, body, opts \\ [])

Makes a POST request.

See request/4 for possible options.

post!(url, body, opts \\ [])

See post/3.

request(method, url, body, opts \\ [])

Makes a HTTP request.

Returns {:ok, response} or {:error, exception}.

Options:

:headers - list of request headers, defaults to [].

:request_middleware - list of middleware to run the request through, defaults to using:

normalize_request_headers/1

default_headers/1

encode/2 with opts

auth/2 with opts[:auth] (if set)

compress/2 with opts[:compress] (if set)

:response_middleware - list of middleware to run the response through, defaults to using:

retry/2 with opts (if retry: true or any of the :retry_* options are set)

decompress/2 with opts

decode/2 with opts

:error_middleware - list of middleware to run the error through, defaults to using:

retry/2 with opts (if retry: true or any of the :retry_* options are set)

:finch - name of the Finch pool to use, defaults to Requests.Finch that is started with the default options. See Finch.start_link/1 for more information.

:pool_timeout - the maximum time to wait for a connection, defaults to 5000. See Finch.request/3 for more information.

:receive_timeout - the maximum time to wait for a response, defaults to 15000. See Finch.request/3 for more information.

Middleware

Requests supports request, response, and error middleware.

A request middleware is any function that accepts and returns a possibly updated Finch.Request struct. An example is default_headers/1.

A response middleware is any function that accepts and returns a possibly updated Finch.Response struct. An example is decompress/2.

An error middleware is any function that accepts and returns a possibly updated exception struct. An example is retry/2.

A response middleware may also return an exception in which case the final return value is switched to {:error, exception} however no further middleware is run on the exception. Similarly, an error middleware may return a response in which case the final return value is switched to {:ok, response} however no further middleware is run on the response.

Notice that some of the built-in middleware functions take more than one argument. In order to use them, you have a couple options:

use capture operator, for example: &Requests.compress(&1, ["gzip"])

use a {module, function, args} tuple, where the first argument (request, response, or exception) will be automatically prepended, for example: {Requests, :compress, ["gzip"]}

Example

opts = [
  request_middleware: [
    &Requests.default_headers/1,
    &IO.inspect(&1, label: :final_request),
  ],
  response_middleware: [
    {IO, :inspect, [[label: :initial_response]]},
    {Requests, :decode, []}
  ]
]

Requests.get!("https://httpbin.org/json", opts)
|> IO.inspect(label: :final_response)

Request middleware

compress(request, algorithms)

Compresses the request body with the given algorithms.

Supported algorithms: :gzip, :deflate, and :identity.

This function also sets the appropriate content-encoding header (unless already set.)

default_headers(request)

Adds default request headers.

Currently these headers are added:

"user-agent" - "requests/0.1.0-dev"

"accept-encoding" - "gzip"

encode(request, opts \\ [])

Encodes the request body based on its shape.

If body is of the following shape, it's encoded and its content-type set accordingly. Otherwise it's unchanged.

Shape	Encoder	Content-Type

{:form, data}	:form_encoder	"application/x-www-form-urlencoded"

{:json, data}	:json_encoder	"application/json"

{:csv, rows}	:csv_encoder	"text/csv"

{:csv, {:stream, rows}}	:csv_stream_encoder	"text/csv"

Options

:form_encoder - defaults to &URI.encode_query/1.

:json_encoder - defaults to &Jason.encode_to_iodata!/1.

:csv_encoder - defaults to &NimbleCSV.RFC4180.dump_to_iodata/1.

:csv_stream_encoder - defaults to &NimbleCSV.RFC4180.dump_to_stream/1.

Examples

iex> Requests.post!("https://httpbin.org/post", {:form, comments: "hello!"}).body["form"]
%{"comments" => "hello!"}

normalize_request_headers(request)

Normalizes request headers.

Turns atom header names into strings, e.g.: :user_agent becomes "user-agent". Non-atom names are returned as is.

Response middleware

decode(response, opts \\ [])

Decodes the response body based on the content-type header.

Options

:json_decoder - if set, used for JSON. Defaults to &Jason.decode/1

:csv_decoder - if set, used for CSV. Defaults to &NimbleCSV.RFC4180.parse_string(&1, skip_headers: false)

:gzip_decoder - used for gzip. Defaults to &:zlib.gunzip/1

:zip_decoder - used for ZIP. Defaults to using :zip.unzip/2

Examples

iex> Requests.get!("https://httpbin.org/json").body
%{...}

iex> Requests.get!("https://httpbin.org/json", json_decoder: fn _ -> "fake" end).body
"fake"

decompress(response, opts \\ [])

Decompresses the response body based on the content-encoding header.

Options

:gzip_decoder - used for the "gzip" and "x-gzip" values. Defaults to &:zlib.gunzip/1

:deflate_decoder - used for the "deflate" value. Defaults to &:zlib.unzip/1

Error middleware

retry(response_or_exception, opts \\ [])

Retries a request on errors.

This function can be used as either or both response and error middleware. It retries a request that resulted in:

response with status 5xx

exception

Options

:retry_max_count - maximum number of retries, defaults to: 2

:retry_delay - sleep this number of milliseconds before making another attempt, defaults to 2000

If all attempts have failed, returns immediately without running any other middleware.

Examples

Server eventually became available:

iex> Requests.get("http://localhost:4000", retry: true)
# Logs:
# 12:37:15.554 [error] Got exception
# ** (Mint.TransportError) connection refused
# Will retry in 2000ms, 2 attempts left
# 12:37:17.557 [error] Got exception
# ** (Mint.TransportError) connection refused
# Will retry in 2000ms, 1 attempt left
{:ok,
 %Finch.Response{
   body: "hello",
   headers: [
     {"cache-control", "max-age=0, private, must-revalidate"},
     {"content-length", "5"},
     {"date", "Mon, 22 Mar 2021 11:37:18 GMT"},
     {"server", "Cowboy"}
   ],
   status: 200
 }}

Server keeps returning 500:

iex> Requests.get("http://localhost:4000", retry: true)
# Logs:
# 12:40:09.186 [error] Got response with status 500
# Headers: [...]
# Will retry in 2000ms, 2 attempts left
# 12:40:11.188 [error] Got response with status 500
# Headers: [...]
# Will retry in 2000ms, 1 attempt left
{:ok,
 %Finch.Response{
   body: "oops",
   headers: [...],
   status: 500
 }}